v 0.1.9 highlight flights arr next hour color coded ETA confirm before changing Key Added Clock Fixed last updated message doesn't change

+++ components/PageView.vue
@@ -6,7 +6,7 @@ defineOptions({
 })
 // const { t } = useI18n()
 // let message =  JSON.parse(localStorage.getItem("listed_flights"))[0]
-const today_date = new Date()
+let today_date = new Date()
 // const date_string = date.toString() // .replace(/T/, ':').replace(/\.\w*/, '')
 const year = today_date.getFullYear().toString()
 const month = `0${today_date.getMonth() + 1}`.slice(-2)
@@ -14,15 +14,8 @@ const day = `0${today_date.getDate().toString()}`.slice(-2)
 const day_number = today_date.getDate()
 const date_formated = `${year}-${month}-${day}`
 const user = useUserStore()
-user.selectedDate = new Date(year,month-1,day_number)
+user.selectedDate = new Date(year, month - 1, day_number)
 user.savedKey = localStorage.getItem('access_key')
-
-
-
-
-
-
-
 
 let filtered_flight = []
 // console.log(localStorage.getItem('tracked_flights'))
@@ -48,7 +41,7 @@ function arr_time() {
   else return new Date(filtered_flight[0].time.scheduled.arrival * 1000).toLocaleTimeString()
 }
 function STA_time() {
-  return new Date(filtered_flight[0].time.scheduled.arrival * 1000).toLocaleString();
+  return new Date(filtered_flight[0].time.scheduled.arrival * 1000).toLocaleTimeString()
 }
 async function go() {
   user.isSortedBySTA = false
@@ -56,6 +49,7 @@ async function go() {
   user.listedFlights = []
   user.message = `last updated: ${today_date.toLocaleString()}`
   localStorage.setItem('last_updated', user.message)
+  console.log(user.message)
   user.trackedFlights.forEach(async (item) => {
     const options = {
       method: 'GET',
@@ -77,21 +71,24 @@ async function go() {
       try {
         user.requestSent = 0
         // console.log(item)
+        // eslint-disable-next-line no-console
         const response = await axios.request(options).then(console.log(item))
         // filtered_flight = await response.data.result.response.data.filter(item => (`0${new Date((item.time.scheduled.arrival) * 1000).getDate().toString()}`).slice(-2) === day)
         filtered_flight = await response.data.result.response.data.filter(item => (new Date((item.time.scheduled.arrival) * 1000).getDate()) === user.selectedDate.getDate())
-        
+
         // console.log(filtered_flight)
         if (filtered_flight[0]) {
+          // eslint-disable-next-line no-async-promise-executor
           await new Promise(async () => {
             user.listedFlights.push({
+              airport_iata: filtered_flight[0].airport.origin.code.iata,
               flight: filtered_flight[0].identification.number.default,
               arr_time: arr_time(),
               // arr_time: filtered_flight[0].time.other.eta,
               status: filtered_flight[0].status.text,
               sta: STA_time(),
               sta_unix: filtered_flight[0].time.scheduled.arrival,
-              eta_unix: arr_time_unix() ,
+              eta_unix: arr_time_unix(),
               aircraft: filtered_flight[0].aircraft.model.text,
             })
             // console.log(filtered_flight)
@@ -99,20 +96,17 @@ async function go() {
             user.requestSent++
             // await new Promise( setTimeout(() => {console.log("wait 2 sec")}, 2000))
           })
-
         }
         // if (user.requestSent == 5 || user.requestSent == 10) {
         //     //wait 2 sec
         //    await new Promise( setTimeout(() => {console.log("wait 1 sec")}, 1000))
         //    await new Promise( setTimeout(() => {console.log("wait 1 sec")}, 1000))
-        //   }  
+        //   }
       }
       catch (error) {
         console.error(error)
       }
-
     }, user.trackedFlights.indexOf(item) * 1000))
-
   })
 
   // message = ''
@@ -128,10 +122,23 @@ function del(index) {
   localStorage.setItem('tracked_flights', user.trackedFlights)
 }
 function save() {
-  user.savedKey = user.inputKey
-  localStorage.setItem('access_key', user.inputKey)
-  user.inputKey = ''
+  if (confirm(`Change API Key to: " ${user.inputKey} " ?`)) {
+    user.savedKey = user.inputKey
+    localStorage.setItem('access_key', user.inputKey)
+    user.inputKey = ''
+  }
 }
+
+function save_key(new_api_key) {
+  if (!new_api_key)
+    new_key = user.inputKey
+  if (confirm(`Change API Key to: " ${new_api_key} " ?`)) {
+    user.savedKey = new_api_key
+    localStorage.setItem('access_key', new_api_key)
+    user.inputKey = ''
+  }
+}
+
 function addFlight() {
   if (user.inputFlight.length > 2) {
     user.trackedFlights.push(user.inputFlight)
@@ -142,16 +149,17 @@ function addFlight() {
 function sortBySTA() {
   // console.log('sortBYSTA pressed')
   if (!user.isSortedBySTA) {
+    // eslint-disable-next-line array-callback-return
     user.listedFlights.sort((a, b) => {
-      if (a.sta_unix < b.sta_unix) {
-        return -1;
-      }
-      if (a.sta_unix > b.sta_unix) {
-        return 1;
-      }
+      if (a.sta_unix < b.sta_unix)
+        return -1
+
+      if (a.sta_unix > b.sta_unix)
+        return 1
     })
-  } else {
-    user.listedFlights.sort().reverse();
+  }
+  else {
+    user.listedFlights.sort().reverse()
   }
   user.isSortedBySTA = !user.isSortedBySTA
   user.isSortedByETA = false
@@ -159,16 +167,17 @@ function sortBySTA() {
 function sortByETA() {
   // console.log('sortBYSTA pressed')
   if (!user.isSortedByETA) {
+    // eslint-disable-next-line array-callback-return
     user.listedFlights.sort((a, b) => {
-      if (a.eta_unix < b.eta_unix) {
-        return -1;
-      }
-      if (a.eta_unix > b.eta_unix) {
-        return 1;
-      }
+      if (a.eta_unix < b.eta_unix)
+        return -1
+
+      if (a.eta_unix > b.eta_unix)
+        return 1
     })
-  } else {
-    user.listedFlights.sort().reverse();
+  }
+  else {
+    user.listedFlights.sort().reverse()
   }
   user.isSortedByETA = !user.isSortedByETA
   user.isSortedBySTA = false
@@ -193,31 +202,42 @@ function showVersionMessage() {
 // const test = respond_1.data.find(({ flight_date }) => flight_date === date_formated)
 // arr_time: new Date(filtered_flight[0].time.estimated.arrival * 1000).toLocaleString()
 // arr_time: `${new Date(filtered_flight[0].time.estimated.arrival * 1000).getHours()}:${new Date(filtered_flight[0].time.estimated.arrival * 1000).getMinutes()}`,
-</script>
-
-
-
-
-
-
 
 function startTime() {
-  const today = new Date();
-  let h = today.getHours();
-  let m = today.getMinutes();
-  let s = today.getSeconds();
-  m = checkTime(m);
-  s = checkTime(s);
-  user.clock_txt =  h + ":" + m + ":" + s;
-  setTimeout(startTime, 1000);
+  today_date = new Date()
+  // console.log(today_date.getTime())
+  const h = today_date.getHours()
+  let m = today_date.getMinutes()
+  let s = today_date.getSeconds()
+  m = checkTime(m)
+  s = checkTime(s)
+  user.clock_txt = `${h}:${m}:${s}`
+  setTimeout(startTime, 1000)
 }
 
 function checkTime(i) {
-  if (i < 10) {i = "0" + i};  // add zero in front of numbers < 10
-  return i;
+  if (i < 10)
+    i = `0${i}` // add zero in front of numbers < 10
+  return i
 }
 
+function etaTxtClass(flight) {
+  if (flight.eta_unix - flight.sta_unix > 3600) // 1 hour
+    return 'color-red-6'
+  if (flight.eta_unix - flight.sta_unix > 1800) // 1/2
+    return 'color-red-4'
+  if ((flight.eta_unix - flight.sta_unix > 900)) // 15 min
+    return 'color-orange'
 
+  if ((flight.eta_unix - flight.sta_unix < -600)) // -10 min
+    return 'color-green-7'
+
+  if ((flight.eta_unix - flight.sta_unix > 600)) // 10 min
+    return ''
+}
+
+startTime()
+</script>
 
 <template>
   <div>
@@ -235,20 +255,32 @@ function checkTime(i) {
     <p>
       <!-- {{ day_number }} -->
     </p>
-      <!-- {{ user.selectedDate }} -->
+    <!-- {{ user.selectedDate }} -->
     <!-- <p> flight tracking page</p> -->
     <p>
       Page's under construction
       <!-- <em text-sm opacity-75>{{ t('intro.desc') }}</em> -->
     </p>
     <div py-4 />
-    <p cursor-pointer text-blue @click="showVersionMessage"> v 0.1.8</p>
-    <div color-black v-show="user.showVersionMessage" style="transition: width 4s;" class="bg-yellow-100 p-4 m-auto w-3/4">
-      <p> </p>
+    <p cursor-pointer text-blue @click="showVersionMessage">
+      v 0.1.9
+    </p>
+    <div
+      v-show="user.showVersionMessage" color-black style="transition: width 4s;"
+      class="bg-yellow-100 p-4 m-auto w-3/4"
+    >
+      <p />
+      <p> v 0.1.9 </p>
+      <p> &nearr; highlight flights arr next hour </p>
+      <p> &nearr; color coded ETA </p>
+      <p> &nearr; confirm before changing Key </p>
+      <p> &nearr; Added Clock </p>
+      <p> &nearr; Fixed last updated message doesn't change </p>
+      <p> v 0.1.8 </p>
       <p> &nearr; Fixed sort by ETA </p>
       <p> &nearr; Fixed message font color </p>
-      <p> v 0.1.7 </p>
-      <p> &nearr; Added select date option </p>
+      <!-- <p> v 0.1.7 </p>
+      <p> &nearr; Added select date option </p> -->
       <!-- <p> v 0.1.6</p>
       <p> &nearr; Added sort by ETA functionality </p>
       <p> &nearr; Changed table format </p>
@@ -262,31 +294,37 @@ function checkTime(i) {
         <!-- m-auto  rounded shadow-md"> -->
         <thead>
           <tr>
+            <th class="border bg-gray">
+              <p i-carbon-departure />
+            </th>
             <th class="border border-gray-300 bg-gray">
               <!-- w-1/3 border border-r-0 border-gray-300 bg-gray p-4 font-normal sm:w-1/4"> -->
               Flight
             </th>
-            <th @click="sortBySTA" class="border bg-gray ">
+            <th class="border bg-gray " @click="sortBySTA">
               STA
-              <p btn i-carbon-arrow-down v-if="user.isSortedBySTA" />
-              <p btn i-carbon-dot-mark v-else />
+              <p v-if="user.isSortedBySTA" btn i-carbon-arrow-down />
+              <p v-else btn i-carbon-dot-mark />
             </th>
-            <th @click="sortByETA" class="border bg-gray ">
+            <th class="border bg-gray " @click="sortByETA">
               ETA
-              <p btn i-carbon-arrow-down v-if="user.isSortedByETA" />
-              <p btn i-carbon-dot-mark v-else />
+              <p v-if="user.isSortedByETA" btn i-carbon-arrow-down />
+              <p v-else btn i-carbon-dot-mark />
             </th>
             <th class="border bg-gray">
               Status
             </th>
-            <th class="border bg-gray">
+            <th v-show="true" class="border bg-gray">
               Aircraft
             </th>
           </tr>
         </thead>
         <!-- use the filtered list -->
         <tbody v-for="flight in user.listedFlights" :key="flight._id">
-          <tr>
+          <tr :class="[{ ['bg-yellow-7']: Math.abs(flight.eta_unix - today_date.getTime() / 1000) < 3600.000 }]">
+            <td class="border">
+              {{ flight.airport_iata }}
+            </td>
             <td class="border">
               <!-- w-1/3 border border-r-0 border-gray-300  p-4 font-normal sm:w-1/4 -->
               {{ flight.flight }}
@@ -294,7 +332,7 @@ function checkTime(i) {
             <td class="border">
               {{ flight.sta }}
             </td>
-            <td class="border">
+            <td :class="etaTxtClass(flight)" class="border">
               {{ flight.arr_time }}
             </td>
             <td class="border">
@@ -307,55 +345,70 @@ function checkTime(i) {
         </tbody>
       </table>
       <div py-4 />
+      <p text-4xl>
+        Clock: {{ user.clock_txt }}
+      </p>
+      <div py-4 />
       <!-- {{ user.listedFlights }} -->
       <!-- {{ filtered_flight }} -->
       {{ user.message }}
       <p>
-
         <select v-model="user.selectedDate">
-          <option disabled="">Choose Date</option>
-          <option :value="new Date(year,month-1,day_number -3)">{{new Date(year,month-1,day_number -3).toLocaleDateString()}}</option>
-          <option :value="new Date(year,month-1,day_number -2)">{{new Date(year,month-1,day_number -2).toLocaleDateString()}}</option>
-          <option :value="new Date(year,month-1,day_number -1)">{{new Date(year,month-1,day_number -1).toLocaleDateString() }}</option>
-          <option :value="new Date(year,month-1,day_number)" >Today</option>
-          <option :value="new Date(year,month-1,day_number +1)">{{new Date(year,month-1,day_number +1).toLocaleDateString()}}</option>
-          <option :value="new Date(year,month-1,day_number +2)">{{new Date(year,month-1,day_number +2).toLocaleDateString()}}</option>
-          <option :value="new Date(year,month-1,day_number +3)">{{new Date(year,month-1,day_number +3).toLocaleDateString()}}</option>
+          <option disabled="">
+            Choose Date
+          </option>
+          <option :value="new Date(year, month - 1, day_number - 3)">
+            {{ new Date(year, month - 1, day_number
+              - 3).toLocaleDateString() }}
+          </option>
+          <option :value="new Date(year, month - 1, day_number - 2)">
+            {{ new Date(year, month - 1, day_number
+              - 2).toLocaleDateString() }}
+          </option>
+          <option :value="new Date(year, month - 1, day_number - 1)">
+            {{ new Date(year, month - 1, day_number
+              - 1).toLocaleDateString() }}
+          </option>
+          <option :value="new Date(year, month - 1, day_number)">
+            Today
+          </option>
+          <option :value="new Date(year, month - 1, day_number + 1)">
+            {{ new Date(year, month - 1, day_number
+              + 1).toLocaleDateString() }}
+          </option>
+          <option :value="new Date(year, month - 1, day_number + 2)">
+            {{ new Date(year, month - 1, day_number
+              + 2).toLocaleDateString() }}
+          </option>
+          <option :value="new Date(year, month - 1, day_number + 3)">
+            {{ new Date(year, month - 1, day_number
+              + 3).toLocaleDateString() }}
+          </option>
         </select>
 
         <button m-3 text-sm btn @click="go">
           Go
         </button>
       </p>
-
-
-
-<p>
-
-    {{ user.clock_txt }}
-
-</p>
-
-
-
-
-
-
     </div>
     <div py-4 />
     <p> Add flight to check status: &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;</p>
     <p>
-      <input id="input2" v-model="user.inputFlight" placeholder=" example EK211 " autocomplete="true" type="text"
-        v-bind="$attrs" p="x-4 y-2" w="250px" text="center" bg="transparent" :border="'~ rounded gray-200 dark:gray-700'"
-        outline="none active:none" @keydown.enter="addFlight" />
+      <input
+        id="input2" v-model="user.inputFlight" placeholder=" example EK211 " autocomplete="true" type="text"
+        v-bind="$attrs" p="x-4 y-2" w="250px" text="center" bg="transparent" border="~ rounded gray-200 dark:gray-700"
+        outline="none active:none" @keydown.enter="addFlight"
+      >
       <button m-3 text-sm btn @click="addFlight">
         Add
       </button>
     </p>
     <p>
-      <input id="input" v-model="user.inputKey" placeholder="enter X-RapidAPI-Key" autocomplete="true" type="text"
-        v-bind="$attrs" p="x-4 y-2" w="250px" text="center" bg="transparent" :border="'~ rounded gray-200 dark:gray-700'"
-        outline="none active:none" @keydown.enter="save" />
+      <input
+        id="input" v-model="user.inputKey" placeholder="enter X-RapidAPI-Key" autocomplete="true" type="text"
+        v-bind="$attrs" p="x-4 y-2" w="250px" text="center" bg="transparent" border="~ rounded gray-200 dark:gray-700"
+        outline="none active:none" @keydown.enter="save"
+      >
       <button m-3 text-sm btn @click="save">
         Save
       </button>
@@ -373,8 +426,10 @@ function checkTime(i) {
     <p id="message">
       user.trackedFlights : click to delete
     </p>
-    <button v-for="flight in user.trackedFlights" :key="flight._id" m-1 bg-red
-      @click="del(user.trackedFlights.indexOf(flight))">
+    <button
+      v-for="flight in user.trackedFlights" :key="flight._id" m-1 bg-red
+      @click="del(user.trackedFlights.indexOf(flight))"
+    >
       {{ flight }}
     </button>
     <div>
@@ -387,11 +442,12 @@ function checkTime(i) {
     </div>
   </div>
 </template>
+
 <style>
-  select {
-    /* display: block; */
+select {
+  /* display: block; */
   /* width: 100%; */
-  padding: .375rem 2.25rem .375rem .75rem;
+  padding: 0.375rem 2.25rem 0.375rem 0.75rem;
   -moz-padding-start: calc(0.75rem - 3px);
   font-size: 1rem;
   font-weight: 400;
@@ -400,17 +456,18 @@ function checkTime(i) {
   background-color: #fff;
   background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3e%3cpath fill='none' stroke='%23343a40' stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='M2 5l6 6 6-6'/%3e%3c/svg%3e");
   background-repeat: no-repeat;
-  background-position: right .75rem center;
+  background-position: right 0.75rem center;
   background-size: 16px 12px;
   border: 1px solid #ced4da;
-  border-radius: .25rem;
-  transition: border-color .15s ease-in-out,box-shadow .15s ease-in-out;
+  border-radius: 0.25rem;
+  transition:
+    border-color 0.15s ease-in-out,
+    box-shadow 0.15s ease-in-out;
   -webkit-appearance: none;
   -moz-appearance: none;
   appearance: none;
   word-wrap: normal;
   margin-bottom: 0;
   word-wrap: normal;
-  }
-
+}
 </style>
